+++ packages/react-composites/src/composites/common/AvatarPersona.tsx
@@ -1,40 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { IPersonaProps, Persona, PersonaInitialsColor } from '@fluentui/react';
-import { AvatarPersonaDataCallback } from '@internal/react-components';
+import { IPersonaProps, Persona } from '@fluentui/react';
+import { AvatarPersonaDataCallback, AvatarPersonaData } from '@internal/react-components';
 import React, { useEffect } from 'react';
-
-/**
- * Custom data attributes for displaying avatar for a user.
- *
- * @public
- */
-export type AvatarPersonaData = {
-  /**
-   * Primary text to display, usually the name of the person.
-   */
-  text?: string;
-  /**
-   * Image URL to use, should be a square aspect ratio and big enough to fit in the image area.
-   */
-  imageUrl?: string;
-  /**
-   * The user's initials to display in the image area when there is no image.
-   * @defaultvalue Derived from `text`
-   */
-  imageInitials?: string;
-  /**
-   * The background color when the user's initials are displayed.
-   * @defaultvalue Derived from `text`
-   */
-  initialsColor?: PersonaInitialsColor | string;
-  /**
-   * The text color when the user's initials are displayed
-   * @defaultvalue `white`
-   */
-  initialsTextColor?: string;
-};
 
 /**
  * @private
