+++ packages/react-composites/src/composites/index.ts
@@ -5,7 +5,6 @@ export type { Disposable, AdapterState } from './common/adapters';
 
 export * from './ChatComposite';
 export * from './CallComposite';
-export type { AvatarPersonaData } from './common/AvatarPersona';
 export { COMPOSITE_ONLY_ICONS, DEFAULT_COMPOSITE_ICONS } from './common/icons';
 export type { CompositeIcons, ChatCompositeIcons, CallCompositeIcons } from './common/icons';
 /* @conditional-compile-remove(call-with-chat-composite) */
